Quick update
I added Javascript for handling the "Next Page," "Previous Page," and the Page Select drop-down. Then I realized there's no UI for it on the "Edit Users" tab. I added these because I started building Thompson Engineering, and was almost able to copy one-for-one the code from Max Foote for this, so figured I'd just include it in the boilerplate. Need to incorporate the UI as well.